# api/public/swagger.yaml
components:
...
    schemas:
...
        MemberType.folder:
            enum:
                - folder
            type: string
        FolderMember:
            properties:
                name:
                    type: string
                type:
                    $ref: '#/components/schemas/MemberType.folder'
                members:
                    items:
                        anyOf:
                            -
                                $ref: '#/components/schemas/FolderMember'
                            -
                                $ref: '#/components/schemas/ServiceMember'
                            -
                                $ref: '#/components/schemas/FileMember'
                    type: array
            required:
                - name
                - type
                - members
            type: object
            additionalProperties: false
        MemberType.service:
            enum:
                - service
            type: string
        ServiceMember:
            properties:
                name:
                    type: string
                type:
                    $ref: '#/components/schemas/MemberType.service'
                code:
                    type: string
            required:
                - name
                - type
                - code
            type: object
            additionalProperties: false
        MemberType.file:
            enum:
                - file
            type: string
        FileMember:
            properties:
                name:
                    type: string
                type:
                    $ref: '#/components/schemas/MemberType.file'
                code:
                    type: string
            required:
                - name
                - type
                - code
            type: object
            additionalProperties: false
        FileTree:
            properties:
                members:
                    items:
                        anyOf:
                            -
                                $ref: '#/components/schemas/FolderMember'
                            -
                                $ref: '#/components/schemas/ServiceMember'
                            -
                                $ref: '#/components/schemas/FileMember'
                    type: array
            required:
                - members
            type: object
            additionalProperties: false
        DeployResponse:
            properties:
                status:
                    type: string
                message:
                    type: string
                streamServiceName:
                    type: string
                example:
                    $ref: '#/components/schemas/FileTree'
            required:
                - status
                - message
            type: object
            additionalProperties: false
        DeployPayload:
            properties:
                appLoc:
                    type: string
                streamWebFolder:
                    type: string
                fileTree:
                    $ref: '#/components/schemas/FileTree'
            required:
                - appLoc
                - fileTree
            type: object
            additionalProperties: false
        UpdateFileResponse:
            properties:
                status:
                    type: string
                message:
                    type: string
            required:
                - status
            type: object
            additionalProperties: false
        TreeNode:
            properties:
                name:
                    type: string
                relativePath:
                    type: string
                absolutePath:
                    type: string
                children:
                    items:
                        $ref: '#/components/schemas/TreeNode'
                    type: array
            required:
                - name
                - relativePath
                - absolutePath
                - children
            type: object
            additionalProperties: false
        GetFileTreeResponse:
            properties:
                status:
                    type: string
                tree:
                    $ref: '#/components/schemas/TreeNode'
            required:
                - status
                - tree
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: api
    version: 0.0.2
    description: 'Api of SASjs server'
    contact:
        name: '4GL Ltd'
openapi: 3.0.0
paths:
...
    /SASjsApi/drive/deploy:
        post:
            operationId: Deploy
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: success, message: 'Files deployed successfully to @sasjs/server.'}
                '400':
                    description: 'Invalid Format'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'Provided not supported data format.'}
                '500':
                    description: 'Execution Error'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'Deployment failed!'}
            summary: 'Creates/updates files within SASjs Drive using provided payload.'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeployPayload'
    /SASjsApi/drive/deploy/upload:
        post:
            operationId: DeployUpload
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: success, message: 'Files deployed successfully to @sasjs/server.'}
                '400':
                    description: 'Invalid Format'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'Provided not supported data format.'}
                '500':
                    description: 'Execution Error'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeployResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'Deployment failed!'}
            description: "Accepts JSON file and zipped compressed JSON file as well.\nCompressed file should only contain one JSON file and should have same name\nas of compressed file e.g. deploy.JSON should be compressed to deploy.JSON.zip\nAny other file or JSON file in zipped will be ignored!"
            summary: 'Creates/updates files within SASjs Drive using uploaded JSON/compressed JSON file.'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                            required:
                                - file
    /SASjsApi/drive/file:
        get:
            operationId: GetFile
            responses:
                '204':
                    description: 'No content'
            summary: 'Get file from SASjs Drive'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: _filePath
                    required: true
                    schema:
                        type: string
                    example: /Public/somefolder/some.file
        delete:
            operationId: DeleteFile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    status: {type: string}
                                required:
                                    - status
                                type: object
            summary: 'Delete file from SASjs Drive'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: _filePath
                    required: true
                    schema:
                        type: string
                    example: /Public/somefolder/some.file
        post:
            operationId: SaveFile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateFileResponse'
                            examples:
                                'Example 1':
                                    value: {status: success}
                '403':
                    description: 'File already exists'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateFileResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'File request failed.'}
            description: "It's optional to either provide `_filePath` in url as query parameter\nOr provide `filePath` in body as form field.\nBut it's required to provide else API will respond with Bad Request."
            summary: 'Create a file in SASjs Drive'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'Location of SAS program'
                    in: query
                    name: _filePath
                    required: false
                    schema:
                        type: string
                    example: /Public/somefolder/some.file.sas
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                filePath:
                                    type: string
                            required:
                                - file
        patch:
            operationId: UpdateFile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateFileResponse'
                            examples:
                                'Example 1':
                                    value: {status: success}
                '403':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateFileResponse'
                            examples:
                                'Example 1':
                                    value: {status: failure, message: 'File request failed.'}
            description: "It's optional to either provide `_filePath` in url as query parameter\nOr provide `filePath` in body as form field.\nBut it's required to provide else API will respond with Bad Request."
            summary: 'Modify a file in SASjs Drive'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'Location of SAS program'
                    in: query
                    name: _filePath
                    required: false
                    schema:
                        type: string
                    example: /Public/somefolder/some.file.sas
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                filePath:
                                    type: string
                            required:
                                - file
    /SASjsApi/drive/folder:
        get:
            operationId: GetFolder
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    folders: {items: {type: string}, type: array}
                                    files: {items: {type: string}, type: array}
                                required:
                                    - folders
                                    - files
                                type: object
            summary: 'Get folder contents from SASjs Drive'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    in: query
                    name: _folderPath
                    required: false
                    schema:
                        type: string
                    example: /Public/somefolder
    /SASjsApi/drive/filetree:
        get:
            operationId: GetFileTree
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetFileTreeResponse'
            summary: 'Fetch file tree within SASjs Drive.'
            tags:
                - Drive
            security:
                -
                    bearerAuth: []
            parameters: []
...
servers:
    -
        url: /
...